compile original module for create-diff-object
Normally, the symbols in the freshly built object file that
create-diff-object used are the same with our original module.

But the name of the symbols may change if use --skip-gcc-check
or the influence of KCFLAGS. This patch compile original module
for create-diff-object, like original vmlinux, to detect this
problem earlier, before we need to insmod klp.ko.

Signed-off-by: Zhou Chengming <zhouchengming1@_huawei.xyz_>